+++ test/agenda/preference_test.exs
@@ -25,7 +25,7 @@ defmodule Agenda.PreferenceTest do
 
   defp talk(name) do
     name
-    |> Agenda.session(lasting: "PT1H")
+    |> Agenda.session(duration: "PT1H")
     |> Session.needs(:room, seats: 100)
   end